+++ phpme/command/phpme_override_method_command.py
@@ -1,6 +1,7 @@
 import sublime_plugin
 import os
 from ..helper import Helper
+from ..utils import Utils
 from ..parser.class_parser import ClassParser
 
 
@@ -51,10 +52,12 @@ class PhpmeOverrideMethodCommand(sublime_plugin.TextCommand):
     def pick_method(self, index):
         method = self.list_methods[index][0]
         namespace = self.list_methods[index][1]
-        if namespace not in self.methods:
-            self.methods[namespace] = {}
-        self.methods[namespace][method] = self.pending[namespace][method]
-        del self.list_methods[index]
+        found = Utils.find(self.pending[namespace], 'name', method)
+        if found:
+            if namespace not in self.methods:
+                self.methods[namespace] = []
+            self.methods[namespace].append(found)
+            del self.list_methods[index]
 
     def on_method_selected(self, index):
         if index > -1:
@@ -88,18 +91,20 @@ class PhpmeOverrideMethodCommand(sublime_plugin.TextCommand):
 
     def find_methods(self, mdef):
         indexes = []
-        use = self.helper.decide_use(mdef['parent']['alias'], mdef['parent']['namespace'], mdef['namespace'], mdef['uses'])
+        use = self.helper.decide_use(mdef['parent']['as'], mdef['parent']['ns'], mdef['namespace'], mdef['uses'])
         symbol = use.split('\\')[-1]
         for namespaces in self.helper.find_symbol(symbol, use):
             if namespaces[0] == use:
                 methods = self.helper.parse_class_tree(None, namespaces[0], namespaces[1], mdef)
-                for namespace, namespace_methods in methods.items():
-                    parent_methods = {}
+                for namespace in sorted(list(methods.keys())):
+                    namespace_methods = methods[namespace]
+                    parent_methods = []
                     for method in namespace_methods:
-                        if method not in mdef['methods'] and method not in indexes and (not self.abstract_only or namespace_methods[method]['abstract']):
-                            indexes.append(method)
-                            self.list_methods.append([method, namespace])
-                            parent_methods[method] = namespace_methods[method]
+                        found = Utils.find(mdef['methods'], 'name', method['name'])
+                        if not found and method['name'] not in indexes and (not self.abstract_only or method['abstract']):
+                            indexes.append(method['name'])
+                            self.list_methods.append([method['name'], namespace])
+                            parent_methods.append(method)
                     if len(parent_methods) > 0:
                         self.pending[namespace] = parent_methods
                 break